refactor: flatten backend structure and remove MinIO dependencies
This commit reorganizes the backend structure for simpler deployment
to Aliyun cloud and removes unused MinIO file serving code.

## Backend Structure Changes

- **Flattened directory structure**: Moved `backend/shared/config_utils.py` → `backend/config_utils.py`
  - Removed `backend/shared/` directory entirely
  - Updated all imports across 7 files (app, index_service, scripts, tests)
  - Simplified import: `from backend.config_utils import load_config`

## Removed Dependencies

- **Removed file_router from main.py and papers.py**:
  - Deleted `get_minio_client()` function (disabled, returns 501)
  - Deleted `file_router` definition and `/files/{bucket}/{key:path}` endpoint (~70 lines)
  - Removed unused imports: `minio.error.S3Error`, `fastapi.responses.Response`
  - Images now served directly from Aliyun OSS (http://oss.paperignition.com/imgs/)

## Nginx Configuration

- **Updated beta_frontend/nginx.conf** for production:
  - Added compatibility routes for `/find_similar`, `/paper_content`, `/get_metadata`
  - These proxy to backend's compatibility endpoints in main.py
  - Aligns with nginx_mac.conf for consistency

## Documentation

- **Updated backend/README.md**: Reflected flattened project structure
- **Updated backend/config_utils.py**: Updated docstring with new location and usage

## Removed Files

- **backend/shared/__init__.py**: No longer needed
- **backend/shared/config_utils.py**: Moved to backend root
- **scripts/imgs2aliyun.py**: Deprecated script

## Other Changes

- **backend/app/routers/users.py**: Commented out deprecated `interests_description` field
- **orchestrator/production_config.yaml**: Disabled `generate_all_papers_blog` stage
- **scripts/launch_test_service.sh**: Uncommented PAPERIGNITION_CONFIG export

## Benefits

1. **Simpler deployment**: 50MB dependencies (backend-only) vs 4-5GB (with Index Service)
2. **Cleaner codebase**: Removed 70+ lines of dead MinIO code
3. **Flatter structure**: Easier to navigate and maintain
4. **Ready for Aliyun**: Production nginx.conf has all necessary routes
