--- backend/shared/__init__.py
@@ -1,12 +1,0 @@
-"""
-Shared utilities for PaperIgnition backend services.
-
-This module contains common functionality shared between:
-- Backend Service (backend/app/)
-
-Last updated: 2026-03-05
-"""
-
-from .config_utils import load_config
-
-__all__ = ['load_config']
