+++ beta_frontend/nginx.conf
@@ -64,6 +64,7 @@ server {
         # Rate limiting
         limit_req zone=api_limit burst=20 nodelay;
 
+
         # Proxy to backend service (localhost)
         proxy_pass http://localhost:8000;
         proxy_set_header Host $host;
@@ -110,6 +111,24 @@ server {
         proxy_connect_timeout 7d;
         proxy_send_timeout 7d;
         proxy_read_timeout 7d;
+    }
+
+    # ======================= Compatibility Routes =======================
+    # Frontend calls these without /api/ prefix - proxy to backend compatibility endpoints
+    location ~ ^/(find_similar|paper_content|get_metadata)/ {
+        # Rate limiting
+        limit_req zone=api_limit burst=20 nodelay;
+
+        proxy_pass http://localhost:8000;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto https;
+
+        # Timeout settings
+        proxy_connect_timeout 60s;
+        proxy_send_timeout 60s;
+        proxy_read_timeout 60s;
     }
 
     # ======================= Static Files =======================
