+++ scripts/list_users.py
@@ -14,7 +14,7 @@ from pathlib import Path
 project_root = Path(__file__).parent.parent
 sys.path.append(str(project_root))
 
-from backend.shared.config_utils import load_config
+from backend.config_utils import load_config
 from backend.app.db_utils import DatabaseManager
 from sqlalchemy import select
 from backend.app.models.users import User
